Fixes #83 and reveals more improvement potential

+++ ref_app/src/app/benchmark/app_benchmark_wide_decimal.cpp
@@ -31,7 +31,7 @@ bool app::benchmark::run_wide_decimal()
 
   using local_allocator_type = util::n_slot_array_allocator<void, local_elem_number, 14U>;
 
-  using dec101_t = math::wide_decimal::decwide_t<wide_decimal_digits10, local_limb_type, local_allocator_type, double>;
+  using dec101_t = math::wide_decimal::decwide_t<wide_decimal_digits10, local_limb_type, local_allocator_type, float>;
 
   const dec101_t s = sqrt(dec101_t(123456U) / 100);
